refactor: put onBadUrl in routing file*
