+++ lib/route.js
@@ -26,7 +26,8 @@ const {
   FST_ERR_ROUTE_METHOD_INVALID,
   FST_ERR_ROUTE_BODY_VALIDATION_SCHEMA_NOT_SUPPORTED,
   FST_ERR_ROUTE_BODY_LIMIT_OPTION_NOT_INT,
-  FST_ERR_HOOK_INVALID_ASYNC_HANDLER
+  FST_ERR_HOOK_INVALID_ASYNC_HANDLER,
+  FST_ERR_BAD_URL
 } = require('./errors')
 
 const {
@@ -51,10 +52,13 @@ const {
 const { buildErrorHandler } = require('./error-handler')
 const { createChildLogger } = require('./logger-factory.js')
 const { getGenReqId } = require('./reqIdGenFactory.js')
+const Reply = require('./reply.js')
+const Request = require('./request.js')
 
 function buildRouting (options) {
   const router = FindMyWay({
     defaultRoute,
+    onBadUrl,
     ...options.config
   })
 
@@ -70,6 +74,8 @@ function buildRouting (options) {
   let return503OnClosing
   let globalExposeHeadRoutes
   let keepAliveConnections
+  let frameworkErrors
+  let onBadUrlContext
 
   let closing = false
 
@@ -92,6 +98,8 @@ function buildRouting (options) {
       ignoreDuplicateSlashes = options.ignoreDuplicateSlashes
       return503OnClosing = Object.hasOwn(options, 'return503OnClosing') ? options.return503OnClosing : true
       keepAliveConnections = fastifyArgs.keepAliveConnections
+      frameworkErrors = fastifyArgs.frameworkErrors
+      onBadUrlContext = fastifyArgs.onBadUrlContext
     },
     routing: router.lookup.bind(router), // router func to find the right handler to call
     route, // configure a route in the fastify instance
@@ -107,17 +115,39 @@ function buildRouting (options) {
   }
 
   // If the router does not match any route, every request will land here
-    // req and res are Node.js core objects
-    function defaultRoute (req, res) {
-      if (req.headers['accept-version'] !== undefined) {
-        // we remove the accept-version header for performance result
-        // because we do not want to go through the constraint checking
-        // the usage of symbol here to prevent any collision on custom header name
-        req.headers[kRequestAcceptVersion] = req.headers['accept-version']
-        req.headers['accept-version'] = undefined
-      }
-      fourOhFour.router.lookup(req, res)
+  // req and res are Node.js core objects
+  function defaultRoute (req, res) {
+    if (req.headers['accept-version'] !== undefined) {
+      // we remove the accept-version header for performance result
+      // because we do not want to go through the constraint checking
+      // the usage of symbol here to prevent any collision on custom header name
+      req.headers[kRequestAcceptVersion] = req.headers['accept-version']
+      req.headers['accept-version'] = undefined
     }
+    fourOhFour.router.lookup(req, res)
+  }
+
+  function onBadUrl (path, req, res) {
+    if (frameworkErrors) {
+      const id = getGenReqId(onBadUrlContext.server, req)
+      const childLogger = createChildLogger(onBadUrlContext, logger, req, id)
+
+      const request = new Request(id, null, req, null, childLogger, onBadUrlContext)
+      const reply = new Reply(res, request, childLogger)
+
+      if (disableRequestLogging === false) {
+        childLogger.info({ req: request }, 'incoming request')
+      }
+
+      return frameworkErrors(new FST_ERR_BAD_URL(path), request, reply)
+    }
+    const body = `{"error":"Bad Request","code":"FST_ERR_BAD_URL","message":"'${path}' is not a valid url component","statusCode":400}`
+    res.writeHead(400, {
+      'Content-Type': 'application/json',
+      'Content-Length': body.length
+    })
+    res.end(body)
+  }
 
   function addConstraintStrategy (strategy) {
     throwIfAlreadyStarted('Cannot add constraint strategy!')
